move library to com.apialerts.client
