+++ library/src/commonMain/kotlin/com/apialerts/client/Constants.kt
@@ -1,4 +1,4 @@
-package com.apialerts
+package com.apialerts.client
 
 const val BASE_URL = "https://api.apialerts.com"
 const val INTEGRATION_NAME = "kotlin"
