+++ library/src/commonMain/kotlin/com/apialerts/client/util/ErrorObject.kt
@@ -1,4 +1,4 @@
-package com.apialerts.util
+package com.apialerts.client.util
 
 internal data class ErrorObject(
     val message: String,
