+++ library/src/commonTest/kotlin/com/apialerts/client/contract/ErrorResponseTest.kt
@@ -1,6 +1,7 @@
-package com.apialerts.contract
+package com.apialerts.client.contract
 
-import com.apialerts.util.networkJson
+import com.apialerts.client.contract.ErrorResponse
+import com.apialerts.client.util.networkJson
 import kotlin.test.Test
 import kotlin.test.assertEquals
