+++ library/src/commonTest/kotlin/com/apialerts/client/contract/EventRequestTest.kt
@@ -1,5 +1,6 @@
-package com.apialerts.contract
+package com.apialerts.client.contract
 
+import com.apialerts.client.contract.EventRequest
 import kotlin.test.Test
 import kotlin.test.assertNotNull
